requirements.txt
airflow==0.6
apache_airflow_providers_http==4.2.0
django_environ==0.9.0
environ==1.0
numpy==1.21.6
office365==0.3.15
Office365_REST_Python_Client==2.4.1
openpyxl==3.1.1
pandas==1.1.5
psycopg2==2.9.6
psycopg2_binary==2.9.5
pyodbc==4.0.0_unsupported
requests==2.28.1
SQLAlchemy==1.4.45
tqdm==4.64.0
python-dotenv==1.0.0
